+++ source/utils/adios_iotest/hdf5Stream.h
@@ -41,7 +41,7 @@ public:
 private:
     MPI_Comm comm;
     hid_t hdf5Type(std::string &type);
-    int nSteps = 0;
+    size_t nSteps = 0;
     void defineHDF5Array(const std::shared_ptr<VariableInfo> ov);
     void putHDF5Array(const std::shared_ptr<VariableInfo> ov, size_t step);
     void getHDF5Array(std::shared_ptr<VariableInfo> ov, size_t step);
