+++ testing/adios2/engine/bp/TestBPStepsFileLocalArray.cpp
@@ -129,7 +129,7 @@ TEST_P(BPStepsFileLocalArrayReaders, EveryStep)
 
         auto var_i32 = io.DefineVariable<int32_t>("i32", shape, start, count);
 
-        for (int step = 0; step < NSteps; ++step)
+        for (int step = 0; step < static_cast<int>(NSteps); ++step)
         {
             // Generate test data for each process uniquely
             m_TestData[step] = GenerateData(step, mpiRank, mpiSize);
@@ -278,7 +278,7 @@ TEST_P(BPStepsFileLocalArrayReaders, NewVarPerStep)
 
         adios2::Engine engine = io.Open(fname, adios2::Mode::Write);
 
-        for (int step = 0; step < NSteps; ++step)
+        for (int step = 0; step < static_cast<int>(NSteps); ++step)
         {
             const std::string varName = lf_VarName(step);
             auto var = io.DefineVariable<int32_t>(varName, shape, start, count);
@@ -317,7 +317,7 @@ TEST_P(BPStepsFileLocalArrayReaders, NewVarPerStep)
                    ", block by block"
                 << std::endl;
         }
-        for (int step = 0; step < NSteps; ++step)
+        for (int step = 0; step < static_cast<int>(NSteps); ++step)
         {
             const std::string varName = lf_VarName(step);
             auto var = io.InquireVariable<int32_t>(varName);
@@ -353,7 +353,7 @@ TEST_P(BPStepsFileLocalArrayReaders, NewVarPerStep)
                 << "Read with Stream reading mode step by step, block by block"
                 << std::endl;
         }
-        for (int step = 0; step < NSteps; ++step)
+        for (int step = 0; step < static_cast<int>(NSteps); ++step)
         {
             engine.BeginStep();
             const std::string varName = lf_VarName(step);
@@ -454,12 +454,12 @@ TEST_P(BPStepsFileLocalArrayParameters, EveryOtherStep)
 
         auto var_i32 = io.DefineVariable<int32_t>("i32", shape, start, count);
         auto var_step = io.DefineVariable<int>("step");
-        for (int step = 0; step < NSteps; ++step)
+        for (int step = 0; step < static_cast<int>(NSteps); ++step)
         {
             // Generate test data for each process uniquely
             engine.BeginStep();
             engine.Put(var_step, step);
-            if (step % 2 == Oddity)
+            if (step % 2 == static_cast<int>(Oddity))
             {
                 m_TestData.push_back(GenerateData(step, mpiRank, mpiSize));
                 std::cout << "Rank " << mpiRank << " write step " << step
@@ -535,7 +535,7 @@ TEST_P(BPStepsFileLocalArrayParameters, EveryOtherStep)
         }
 
         size_t writtenStep = 0;
-        for (int step = 0; step < NSteps; ++step)
+        for (size_t step = 0; step < NSteps; ++step)
         {
             engine.BeginStep();
             if (step % 2 == Oddity)
